# [TYPE] Chapter & Stage Data도입
# 본문은 아래에 작성


# --- COMMIT END ---
# [TYPE] 리스트
# ADD : 새로운 파일 추가
# FEAT : 기능 추가 및 수정
# DESIGN : 디자인 추가 및 수정
# FIX : 버그 수정
# DOCS : 문서 (README 등)
# TEST : 테스트 코드
# REFACTOR : 코드 리팩토링 (성능 개선)
# COMMENT : 주석 추가
# CHORE : 빌드 및 설정 파일 관리
# ------------------
#     타입 이름은 전체 대문자로
#     제목은 명령문으로
#     제목 끝에 마침표(.) 금지
#     본문에 여러줄의 메시지를 작성할 땐 "-"로 구분
# ------------------

+++ DeliveryRun/DeliveryRun iOS/Model/Stage.swift
@@ -11,8 +11,8 @@ import Foundation
 class Stage: Codable{
     let name: String
     let image: String
-    let targetRecord: Double
-    var myRecord: Double
+    let targetRecord: Float
+    var myRecord: Float
     var star: Int {
         get {
             if myRecord == 0.0 { return 0 }
@@ -24,7 +24,7 @@ class Stage: Codable{
     }
     var isLock: Bool
     
-    init(name: String, image: String, targetRecord: Double, myRecord:Double,isLock: Bool) {
+    init(name: String, image: String, targetRecord: Float, myRecord:Float,isLock: Bool) {
         self.name = name
         self.image = image
         self.targetRecord = targetRecord
